requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
Django==3.2.8
django-cors-headers==3.10.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
idna==3.2
PyJWT==2.3.0
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7
